Fixed crash when restarting ChartbeatService
We are seeing a crash
```
Attempt to read from field 'android.os.MessageQueue android.os.Looper.mQueue' on a null object reference
  at android.os.Handler.<init>(Handler.java:229)
  at android.os.Handler.<init>(Handler.java:137)
  at com.chartbeat.androidsdk.ChartbeatServiceHandler.<init>(Unknown Source)
  at com.chartbeat.androidsdk.ChartbeatService.onCreate(Unknown Source)
```
which seems to be cause by `onDestroy()` followed by `onCreate()` being called on the service without the process dying. This mean `bThread` is not recreated and `getLooper()` is null. By making the fields non-static and nulling out `bgThread` it ensures the thread is properly created again in `onCreate()`.